separating concerns for iframe recipe crafting
