+++ typescript/packages/common-charm/src/types.d.ts
@@ -2,3 +2,8 @@ declare module "*.html?raw" {
   const content: string;
   export default content;
 }
+
+declare module "*.md?raw" {
+  const content: string;
+  export default content;
+}
